warl0k_OT_final2/warl0k_proj_policy_qurant/requirements.txt
numpy
cryptography
paho-mqtt
